feat: create SandyTrap and Defense class
