+++ src/main/java/edu/fiuba/algo3/modelo/defense/SandyTrap.java
@@ -1,0 +1,4 @@
+package edu.fiuba.algo3.modelo.defense;
+
+public class SandyTrap {
+}
